+++ src/core/entities/source.js
@@ -1,6 +1,5 @@
 'use strict';
 
-const Promise = require('bluebird');
 const _ = require('lodash');
 
 const utils = require('../utils');