+++ src/fractal/cli/console.js
@@ -1,6 +1,5 @@
 'use strict';
 
-const Promise = require('bluebird');
 const _ = require('lodash');
 const Table = require('cli-table3');
 const Theme = require('./themes/default');